Fix eventhandler callback bug + rework Data type to copy on creation
Eventhandler callback bug
-------------------------

The eventhandler callback wasn't supposed to return an unsigned int, but a normal int.
Not passing a normal int with a 0 value caused the SMTP idle job to hang forever
meaning that no messages were sent at all.

Copy values on creation of the Data type when receiving events
--------------------------------------------------------------

Since we have no influence on the moment that DC will free() or unref the data that is
passed with events, it is safer to just immediately copy that data into go types and store
it in our Data struct with the event. This allowed the removal of the (kind of awkward)
rawEvent type from events.go.

Logging has also been improved in this commit, unhandled events will be logged with their
respective name when received, using the generated eventNames map.